fix(OutputView): 로또 결과를 제대로 출력하지 못하는 오류 수정

+++ src/main/java/lotto/service/LottoService.java
@@ -43,9 +43,9 @@ public class LottoService {
     }
 
     private double calculateProfitRate(Map<LottoRanking, Integer> winningCounts, int originalPrice) {
-        int prizeMoney = 0;
+        long prizeMoney = 0;
         for (LottoRanking lottoRanking : winningCounts.keySet()) {
-            prizeMoney += lottoRanking.getPrizeMoney() * winningCounts.get(lottoRanking);
+            prizeMoney += (long) lottoRanking.getPrizeMoney() * winningCounts.get(lottoRanking);
         }
         return (double) prizeMoney / originalPrice;
     }
